backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
pgvector>=0.3.0
pydantic-settings>=2.0.0
redis>=5.0.0
celery>=5.4.0
httpx>=0.27.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.0
praw>=7.8.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
